fix(quality): address all code review findings across Python and Go
Python:
- Extract BaseEloExtractor mixin to eliminate duplicated _days_since_last_game
  and _recent_form across all four sport feature extractors
- Add PoissonModel guards: reject rho > 0, clamp score matrix, guard zero sum
- Add EnsembleModel guard: raise on mixing draw/no-draw capable models
- Replace hasattr duck-typing with isinstance(model, CalibratableModel)
- Consolidate _outcome_label into modeling/types.py; remove duplication
- Add structured logging to all bare except-continue blocks in BacktestPipeline
- Add relative imports and minimum-sample guard in tracking/metrics.py
- Raise ValueError on unknown model/sport in CLI instead of silent fallthrough
- Validate paper-trade URL scheme to localhost/127.0.0.1 only
- Wrap output file writes in try/except OSError in CLI commands
- Use Path.open() in CSV loader; wrap parse errors with row-indexed ValueError

Go:
- Fix TOCTOU race in SafeBookmakerClient: reserve dedup key before unlock,
  retract on inner client failure
- Add negative/zero input guards to Devig; skip nil snapshots in DevigLines
- Change BetID from float64 to json.Number to preserve large integer precision
- Change ComputeCLV to return (CLVResult, error) on zero/negative odds
- Fix context goroutine leaks: defer cancel() in paper-trade and market-data
- Add structured logging to httpapi Handler and cachestore Store
- Change oddsAPIOutcome.Price from int to float64 for decimal odds support
- Fix httpClient() to return a 10s-timeout client instead of http.DefaultClient
- Improve kafkapub Close() doc comment accuracy

+++ internal/marketdata/theoddsapi.go
@@ -83,7 +83,7 @@ type oddsAPIMarket struct {
 
 type oddsAPIOutcome struct {
 	Name  string   `json:"name"`
-	Price int      `json:"price"`
+	Price float64  `json:"price"`
 	Point *float64 `json:"point,omitempty"`
 }
 
@@ -263,7 +263,7 @@ func (p *TheOddsAPIProvider) refresh(ctx context.Context, sport Sport, sportKey 
 				for _, outcome := range mkt.Outcomes {
 					side := outcomeNameToSide(outcome.Name, ae.HomeTeam, ae.AwayTeam)
 					lineID := marketID + "_" + bm.Key + "_" + string(side)
-					decimal, err := AmericanOdds(outcome.Price).ToDecimal()
+					decimal, err := AmericanOdds(int(outcome.Price)).ToDecimal()
 					if err != nil {
 						continue
 					}
@@ -277,7 +277,7 @@ func (p *TheOddsAPIProvider) refresh(ctx context.Context, sport Sport, sportKey 
 						BookID:         bm.Key,
 						Side:           side,
 						Label:          outcome.Name,
-						AmericanOdds:   outcome.Price,
+						AmericanOdds:   int(outcome.Price),
 						DecimalOdds:    float64(decimal),
 						RawImpliedProb: float64(rawProb),
 						RecordedAt:     mkt.LastUpdate,
